Refactor Daytona usage proxy runtime wiring
