+++ src/benchflow/providers/sandbox_usage_proxy.py
@@ -26,239 +26,15 @@ from benchflow.trajectories.types import Trajectory
 logger = logging.getLogger(__name__)
 
 _RUNTIME_ROOT = "/tmp/benchflow-usage-proxy"
-_NODE_PROXY_SOURCE = r"""#!/usr/bin/env node
-const fs = require("fs");
-const http = require("http");
-const https = require("https");
-const { URL } = require("url");
 
-function getArg(name) {
-  const prefix = `--${name}=`;
-  const arg = process.argv.find((value) => value.startsWith(prefix));
-  return arg ? arg.slice(prefix.length) : "";
-}
 
-function getConfig(name, argName) {
-  return process.env[`BENCHFLOW_USAGE_PROXY_${name}`] || getArg(argName);
-}
+def _read_node_proxy_source() -> str:
+    return (
+        Path(__file__).with_name("assets") / "sandbox_usage_proxy.js"
+    ).read_text()
 
-const target = new URL(getConfig("TARGET", "target").replace(/\/+$/, ""));
-const statePath = getConfig("STATE_PATH", "state");
-const logPath = getConfig("LOG_PATH", "log");
-const pidPath = getConfig("PID_PATH", "pid");
-const sessionId = getConfig("SESSION_ID", "session-id");
-const agentName = getConfig("AGENT_NAME", "agent-name");
-const promptCacheRetention = getConfig("PROMPT_CACHE_RETENTION", "prompt-cache-retention");
 
-function sanitizeHeaders(headers) {
-  const result = { ...headers };
-  for (const key of Object.keys(result)) {
-    const lower = key.toLowerCase();
-    if (["connection", "keep-alive", "proxy-authenticate", "proxy-authorization", "te", "trailer", "upgrade"].includes(lower)) {
-      delete result[key];
-    }
-  }
-  return result;
-}
-
-function responseHeaders(headers) {
-  const result = sanitizeHeaders(headers);
-  delete result["content-length"];
-  delete result["Content-Length"];
-  delete result["transfer-encoding"];
-  delete result["Transfer-Encoding"];
-  result["connection"] = "close";
-  return result;
-}
-
-const sensitiveHeaderNames = new Set([
-  "authorization",
-  "proxy-authorization",
-  "x-api-key",
-  "api-key",
-  "openai-api-key",
-  "anthropic-api-key",
-  "x-goog-api-key",
-  "cookie",
-  "set-cookie",
-]);
-const sensitiveQueryNames = new Set([
-  "key",
-  "api_key",
-  "apikey",
-  "access_token",
-]);
-
-function captureHeaders(headers) {
-  const result = { ...headers };
-  for (const key of Object.keys(result)) {
-    if (sensitiveHeaderNames.has(key.toLowerCase())) {
-      result[key] = "__BENCHFLOW_REDACTED__";
-    }
-  }
-  return result;
-}
-
-function capturePath(requestUrl) {
-  const parsed = new URL(requestUrl, "http://benchflow.local");
-  for (const key of Array.from(parsed.searchParams.keys())) {
-    if (sensitiveQueryNames.has(key.toLowerCase())) {
-      parsed.searchParams.set(key, "__BENCHFLOW_REDACTED__");
-    }
-  }
-  return `${parsed.pathname}${parsed.search}`;
-}
-
-function appendCapture(record) {
-  fs.appendFileSync(logPath, JSON.stringify(record) + "\n", { encoding: "utf8" });
-}
-
-function bodyB64(chunks) {
-  return Buffer.concat(chunks).toString("base64");
-}
-
-function upstreamPath(requestUrl) {
-  const basePath = target.pathname.replace(/\/+$/, "");
-  if (!basePath) return requestUrl;
-  return `${basePath}${requestUrl.startsWith("/") ? requestUrl : `/${requestUrl}`}`;
-}
-
-function maybeApplyPromptCacheRetention(requestUrl, headers, body) {
-  if (!promptCacheRetention) return { headers, body };
-  const requestPath = new URL(requestUrl, "http://127.0.0.1").pathname.replace(/\/+$/, "");
-  if (!requestPath.endsWith("/responses") && !requestPath.endsWith("/chat/completions")) {
-    return { headers, body };
-  }
-  const contentEncoding = String(headers["content-encoding"] || headers["Content-Encoding"] || "identity").toLowerCase();
-  if (contentEncoding !== "identity") return { headers, body };
-  try {
-    const parsed = body.length > 0 ? JSON.parse(body.toString("utf8")) : {};
-    if (typeof parsed !== "object" || parsed === null || Array.isArray(parsed)) {
-      return { headers, body };
-    }
-    if (Object.prototype.hasOwnProperty.call(parsed, "prompt_cache_retention")) {
-      return { headers, body };
-    }
-    parsed.prompt_cache_retention = promptCacheRetention;
-    const updatedBody = Buffer.from(JSON.stringify(parsed));
-    const updatedHeaders = { ...headers };
-    delete updatedHeaders["content-encoding"];
-    delete updatedHeaders["Content-Encoding"];
-    updatedHeaders["content-type"] = updatedHeaders["content-type"] || updatedHeaders["Content-Type"] || "application/json";
-    updatedHeaders["content-length"] = updatedBody.length;
-    return { headers: updatedHeaders, body: updatedBody };
-  } catch (_error) {
-    return { headers, body };
-  }
-}
-
-const server = http.createServer((clientReq, clientRes) => {
-  const healthPath = new URL(clientReq.url, "http://127.0.0.1").pathname;
-  if ((clientReq.method === "GET" || clientReq.method === "HEAD") && ["/health", "/__benchflow_health"].includes(healthPath.replace(/\/+$/, "") || "/")) {
-    const payload = Buffer.from(JSON.stringify({ status: "ok" }));
-    clientRes.writeHead(200, {
-      "content-type": "application/json",
-      "content-length": clientReq.method === "HEAD" ? 0 : payload.length,
-    });
-    if (clientReq.method !== "HEAD") clientRes.end(payload);
-    else clientRes.end();
-    return;
-  }
-
-  const requestChunks = [];
-  clientReq.on("data", (chunk) => requestChunks.push(chunk));
-  clientReq.on("end", () => {
-    const originalRequestBody = Buffer.concat(requestChunks);
-    const startedAt = Date.now();
-    const prepared = maybeApplyPromptCacheRetention(
-      clientReq.url,
-      sanitizeHeaders(clientReq.headers),
-      originalRequestBody,
-    );
-    const requestBody = prepared.body;
-    const upstreamHeaders = prepared.headers;
-    upstreamHeaders.host = target.host;
-    if (requestBody.length > 0) upstreamHeaders["content-length"] = requestBody.length;
-
-    const options = {
-      protocol: target.protocol,
-      hostname: target.hostname,
-      port: target.port || (target.protocol === "https:" ? 443 : 80),
-      method: clientReq.method,
-      path: upstreamPath(clientReq.url),
-      headers: upstreamHeaders,
-    };
-    const transport = target.protocol === "https:" ? https : http;
-    const upstreamReq = transport.request(options, (upstreamRes) => {
-      const responseChunks = [];
-      clientRes.writeHead(
-        upstreamRes.statusCode || 502,
-        upstreamRes.statusMessage || "OK",
-        responseHeaders(upstreamRes.headers),
-      );
-      upstreamRes.on("data", (chunk) => {
-        responseChunks.push(chunk);
-        clientRes.write(chunk);
-      });
-      upstreamRes.on("end", () => {
-        clientRes.end();
-        appendCapture({
-          session_id: sessionId,
-          agent_name: agentName,
-          duration_ms: Date.now() - startedAt,
-          request: {
-            method: clientReq.method,
-            path: capturePath(clientReq.url),
-            headers: captureHeaders(clientReq.headers),
-            body_b64: requestBody.toString("base64"),
-          },
-          response: {
-            status_code: upstreamRes.statusCode || 0,
-            headers: captureHeaders(upstreamRes.headers),
-            body_b64: bodyB64(responseChunks),
-          },
-        });
-      });
-    });
-    upstreamReq.on("error", (error) => {
-      const payload = Buffer.from(JSON.stringify({ error: String(error.message || error) }));
-      clientRes.writeHead(502, {
-        "content-type": "application/json",
-        "content-length": payload.length,
-        "connection": "close",
-      });
-      clientRes.end(payload);
-      appendCapture({
-        session_id: sessionId,
-        agent_name: agentName,
-        duration_ms: Date.now() - startedAt,
-        request: {
-          method: clientReq.method,
-          path: capturePath(clientReq.url),
-          headers: captureHeaders(clientReq.headers),
-          body_b64: requestBody.toString("base64"),
-        },
-        response: {
-          status_code: 502,
-          headers: { "content-type": "application/json" },
-          body_b64: payload.toString("base64"),
-        },
-      });
-    });
-    if (requestBody.length > 0) upstreamReq.write(requestBody);
-    upstreamReq.end();
-  });
-});
-
-server.listen(0, "127.0.0.1", () => {
-  const address = server.address();
-  fs.writeFileSync(pidPath, String(process.pid));
-  fs.writeFileSync(statePath, JSON.stringify({ port: address.port, pid: process.pid }));
-});
-
-process.on("SIGTERM", () => server.close(() => process.exit(0)));
-"""
-
+_NODE_PROXY_SOURCE = _read_node_proxy_source()
 _NODE_LAUNCHER_SOURCE = r"""
 const fs = require("fs");
 const { spawn } = require("child_process");
